Save Functionality Started
Added the .ser file to save objects. Serialized both the Agent class and the Note class. Added two methods, save which saves the current objects stored in the arraylist, and initialize which creates the arraylist based on the .ser file. Having issues with an IO exception being thrown whenever an agent object attempts to be saved with a note object.